[FE] Refactor: 어플리케이션 성능 최적화
- Account Router 코드 스플릿팅
