+++ fe/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^16.9.8",
     "@types/styled-components": "^5.1.4",
     "axios": "^0.21.0",
+    "cra-bundle-analyzer": "^0.1.0",
     "dayjs": "^1.9.6",
     "http-proxy-middleware": "^1.0.6",
     "mobx": "^6.0.4",
@@ -26,7 +27,8 @@
     "socket.io-client": "^4.0.0",
     "styled-components": "^5.2.1",
     "styled-reset": "^4.3.1",
-    "typescript": "~4.0.5"
+    "typescript": "~4.0.5",
+    "yarn": "^1.22.10"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -93,6 +95,7 @@
     "lint-staged": "^10.5.3",
     "prettier": "^2.1.2",
     "react-docgen-typescript-loader": "^3.7.2",
-    "react-is": "^17.0.1"
+    "react-is": "^17.0.1",
+    "webpack-bundle-analyzer": "^4.4.2"
   }
 }
