+++ fe/src/App.tsx
@@ -7,8 +7,6 @@ import {
   Redirect,
 } from 'react-router-dom';
 import GlobalThemeProvider from 'styles/GlobalThemeProvider';
-import AccountListPage from 'pages/AccountListPage';
-import AccountUpdatePage from 'pages/AccountUpdatePage';
 import CalendarPage from 'pages/CalendarPage';
 import AuthCheck from 'pages/AuthCheck';
 import StatisticsPage from 'pages/StatisticsPage';
@@ -85,18 +83,7 @@ const LoginRouter = () => {
   );
 };
 
-const AccountRouter = () => {
-  const { url } = useRouteMatch();
-  return (
-    <>
-      <AuthCheck />
-      <Switch>
-        <Route path={`${url}/update`} component={AccountUpdatePage} />
-        <Route path={`${url}`} component={AccountListPage} />
-      </Switch>
-    </>
-  );
-};
+const AccountRouter = React.lazy(() => import('routes/Account'));
 
 const App = () => {
   return (
